production-cms-backend-advanced/requirements.txt
Django==5.0.6
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
django-environ==0.11.2
psycopg2-binary==2.9.9
Pillow==10.3.0
drf-spectacular==0.27.0
django-ckeditor==6.7.0 # For rich text content editing
whitenoise==6.6.0 # For serving static files in production
gunicorn==22.0.0 # WSGI HTTP Server
celery==5.3.6 # For async tasks (example, not fully integrated in core code for brevity)
redis==5.0.6 # For caching and Celery broker
django-redis==5.4.0 # Django cache backend for Redis
faker==24.4.0 # For seeding data
pytest==8.2.2 # Testing
pytest-django==4.8.0 # Testing
pytest-cov==5.0.0 # Coverage
ipdb==0.13.13 # Debugging